requirements.txt
streamlit
pandas
plotly
folium
streamlit-folium
fpdf
python-docx